requirements.txt
Jinja2==3.0.2
Markdown==3.2.1
PyYAML==5.4.1